refactor: begin to replace final-form
