+++ web/src/components/inputs/index.ts
@@ -1,6 +1,5 @@
 export { default as TextField } from "./TextField";
 export { default as TextFieldWide } from "./TextFieldWide";
-export { default as PasswordField } from "./PasswordField";
 export { default as TextAreaWide } from "./TextAreaWide";
 export { default as MultiSelectField } from "./MultiSelectField";
 export { default as RadioFieldset } from "./RadioFieldset";
