--- web/src/screens/filters/inputs/index.ts
@@ -1,7 +1,0 @@
-export { default as DownloadClientSelect } from "./DownloadClientSelect";
-export { default as TextField } from "./TextField";
-export { default as Select } from "./Select";
-export { default as SwitchGroup } from "./SwitchGroup";
-export { default as MultiSelect } from "./MultiSelect";
-export { default as NumberField } from "./NumberField";
-export { Switch } from "./Switch";
